Coding Examples/API Example Python/requirements.txt
requests
numpy
pandas
python-dotenv
pyyaml